+++ src/main/java/com/back/csaback/Services/EvaluationRubriqueService.java
@@ -57,14 +57,14 @@ public class EvaluationRubriqueService {
 
 
     public void detachRubriqueFromEval(Integer idRubriqueEvaluation){
-        Optional<RubriqueEvaluation> rubriqueEvaluation = rubriqueEvaluationRepository.findById(idRubriqueEvaluation.longValue());
+        Optional<RubriqueEvaluation> rubriqueEvaluation = rubriqueEvaluationRepository.findById(idRubriqueEvaluation);
         if(rubriqueEvaluation.isEmpty()) throw new EntityNotFoundException();
         List<RubriqueEvaluation> rubriques = rubriqueEvaluationRepository.findByEvaluation(rubriqueEvaluation.get().getIdEvaluation());
         rubriques.sort(Comparator.comparingInt(RubriqueEvaluation::getOrdre));
         String etat = rubriqueEvaluation.get().getIdEvaluation().getEtat();
         if (!etat.equals("ELA")) throw new IllegalStateException();
         rubriques.remove(rubriqueEvaluation.get());
-        rubriqueEvaluationRepository.deleteById(idRubriqueEvaluation.longValue());
+        rubriqueEvaluationRepository.deleteById(idRubriqueEvaluation);
 
         for(RubriqueEvaluation rubrique : rubriques){
             rubrique.setOrdre((short) (rubriques.indexOf(rubrique)+1));
